refactor(oidc): fold the admin key panel into rotation
The panel is a projection of the rotation state machine (which gate
blocks rotate/revoke right now), and no other admin page has a
dedicated view-model file — merge panel.rs into rotation.rs as its
final section. The gate helpers return to being private, and the
pub(crate) panel re-export now comes from rotation.

Also delete two comments orphaned by the split: the async-tests section
header at the bottom of mod.rs and the redundant panel banner.
